requirements.txt
Django~=3.2.2
djangorestframework ~=3.15.2
drf-yasg ~= 1.20.0

psycopg2-binary~=2.8.6
